размер окошка вновь сохряняется =) начал исправлять long на int в целочисленных опциях
git-svn-id: https://only.mawhrin.net/repos/FBReaderJ/trunk@245 6a642e6f-84f6-412e-ac94-c4a38d5a04b0

+++ src/org/zlibrary/core/options/ZLIntegerOption.java
@@ -9,26 +9,26 @@ import org.zlibrary.core.options.util.ZLToStringConverter;
  *
  */
 public final class ZLIntegerOption extends ZLOption{
-	private long myValue;
-	private final long myDefaultValue;
+	private int myValue;
+	private final int myDefaultValue;
 	
-	public ZLIntegerOption (String category, String group, String optionName, long defaultValue) {
+	public ZLIntegerOption (String category, String group, String optionName, int defaultValue) {
 		super(category, group, optionName);
 		myDefaultValue = defaultValue;
 		myValue = myDefaultValue;
 	}
 	
-	public long getValue() {
+	public int getValue() {
 		if (!myIsSynchronized) {
 			String strDefaultValue = ZLToStringConverter.convert(myDefaultValue);
 			String value = myConfig.getValue(myGroup, myOptionName, strDefaultValue);
-			myValue = ZLFromStringConverter.getLongValue(value);
+			myValue = ZLFromStringConverter.getIntegerValue(value);
 			myIsSynchronized = true;
 		}
 		return myValue;
 	}
 	
-	public void setValue(long value) {
+	public void setValue(int value) {
 		if (myIsSynchronized && (myValue == value)) {
 			return;
 		}
